+++ semana9/intro-redux/app/src/components/TaskList.js
@@ -1,107 +1,33 @@
 import React from 'react'
-
 import { connect } from 'react-redux'
-import { removeTask, editTask, markTaskAsComplete } from '../actions'
 
-import TextField from '@material-ui/core/TextField';
+import TaskItem from './TaskItem';
+
 import List from '@material-ui/core/List';
-import ListItem from '@material-ui/core/ListItem';
-import ListItemIcon from '@material-ui/core/ListItemIcon';
-import ListItemText from '@material-ui/core/ListItemText';
-import Checkbox from '@material-ui/core/Checkbox';
-import Button from '@material-ui/core/Button';
-import ListItemSecondaryAction from '@material-ui/core/ListItemSecondaryAction';
-import IconButton from '@material-ui/core/IconButton';
-import DeleteIcon from '@material-ui/icons/Delete';
 
+function TaskList(props) {
+	const { tasks, filter } = props
+	// console.log(tasks)
+	// console.log(filter)
 
-class TaskList extends React.Component {
-	constructor(props) {
-		super(props)
-		this.state = {
-			editingId: '',
-			editInput: '',
-			inputHere: false
-		}
-	}
-
-	openEditTask = (id, text) => {
-		this.setState({
-			editingId: id,
-			editInput: text,
-			inputHere: true
-		})
-	}
-
-	inputEditChange = (event) => {
-		this.setState({ editInput: event.target.value })
-	}
-
-	disappearsInput = (id, input) => {
-		this.props.onEditTask(id, input)
-		this.setState({ inputHere: false })
-	}
-
-	render() {
-		const { tasks, onDelete, onMarkTaskAsComplete } = this.props
-		console.log(tasks)
-		const filter = this.props.filter
-		console.log(filter)
-
-		return (
-			<List>
-				{tasks
-					.filter((task) => {
-						// if(filter === 'todas') return true
-						if(filter === 'pendentes') {
-							return task.completed === false
-						} 
-						if(filter === 'completas') {
-							return task.completed === true
-						} 
-						return true
-					})
-					.map(task => (
-					<ListItem key={task.id} button>
-
-						<ListItemIcon>
-							<Checkbox edge="start"
-								checked={task.completed}
-								onChange={() => onMarkTaskAsComplete(task.id)} />
-						</ListItemIcon>
-
-						{this.state.editingId === task.id && this.state.inputHere
-							? (<>
-								<TextField id="standard-basic" label="Editar"
-									value={this.state.editInput}
-									onChange={this.inputEditChange}>
-								</TextField>
-								<ListItemIcon>
-									<Button color='primary'
-										onClick={() => this.disappearsInput(task.id, this.state.editInput)}>
-										SALVAR
-										</Button>
-								</ListItemIcon>
-							</>) : (
-								<ListItemText 
-									onClick={() => this.openEditTask(task.id, task.text)}>
-										{task.completed ? <strike>{task.text}</strike> : <span>{task.text}</span>}
-								</ListItemText>
-							)
-						}
-
-						<ListItemSecondaryAction>
-							<IconButton edge="end" aria-label="delete"
-								onClick={() => onDelete(task.id)}>
-								<DeleteIcon />
-							</IconButton>
-						</ListItemSecondaryAction>
-
-					</ListItem>
+	return (
+		<List>
+			{tasks
+				.filter((task) => {
+					// if(filter === 'todas') return true
+					if (filter === 'pendentes') {
+						return task.completed === false
+					}
+					if (filter === 'completas') {
+						return task.completed === true
+					}
+					return true
+				})
+				.map(task => (
+					<TaskItem task={task} />
 				))}
-			</List>
-		)
-	}
+		</List>
+	)
 }
 
 const mapStateToProps = (state) => ({
@@ -109,12 +35,4 @@ const mapStateToProps = (state) => ({
 	filter: state.filterReducer
 })
 
-const mapDispatchToProps = dispatch => {
-	return {
-		onDelete: id => dispatch(removeTask(id)),
-		onEditTask: (id, text) => dispatch(editTask(id, text)),
-		onMarkTaskAsComplete: id => dispatch(markTaskAsComplete(id)),
-	}
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(TaskList);
+export default connect(mapStateToProps)(TaskList);
