+++ semana9/intro-redux/app/src/index.css
@@ -11,35 +11,3 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
-
-.centraliza-texto{
-  text-align: center;
-}
-
-.vermelho{
-  color: rgb(146, 1, 1);
-}
-
-.enorme{
-  font-size: 5rem;
-}
-
-.paper{
-  padding: 1rem;
-}
-
-.topo{
-  margin-bottom: 1rem;
-}
-
-.rodape{
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-around;
-  margin-top: 1rem;
-  align-items: center;
-}
-
-.botao{
-  margin: 0.5rem;
-}
